# docs/openapi.json
{
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/timesheet" : {
      "get" : {
        "operationId" : "getTimesheetForUsers",
        "parameters" : [ {
          "in" : "query",
          "name" : "user_ids",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "start_date",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "end_date",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "scope_path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Timesheet"
                }
              }
            },
            "description" : "default response"
          }
        },
        "summary" : "Returns timesheet report for several users for a given period",
        "tags" : [ "Timesheet reports" ]
      }
    },
...
    "/api/users/{user_id}/timesheet" : {
      "get" : {
        "operationId" : "getTimesheet",
        "parameters" : [ {
          "in" : "path",
          "name" : "user_id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "start_date",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "end_date",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "scope_path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Timesheet"
                }
              }
            },
            "description" : "default response"
          }
        },
        "summary" : "Returns timesheet report for a user for a given period",
        "tags" : [ "Timesheet reports" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Project" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "Timesheet" : {
        "type" : "object",
        "properties" : {
          "finishDate" : {
            "type" : "string"
          },
          "startDate" : {
            "type" : "string"
          },
          "totalHours" : {
            "type" : "number",
            "format" : "double"
          },
          "workRecords" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WorkRecord"
            }
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "WorkItem" : {
        "type" : "object",
        "properties" : {
          "html" : {
            "type" : "string"
          },
          "iconUrl" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "project" : {
            "$ref" : "#/components/schemas/Project"
          },
          "title" : {
            "type" : "string"
          }
        }
      },
      "WorkRecord" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string"
          },
          "hours" : {
            "type" : "number",
            "format" : "double"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "workItem" : {
            "$ref" : "#/components/schemas/WorkItem"
          }
        }
      }
    }
  }
}